Clean up app component
