+++ explorer/src/App.tsx
@@ -1,20 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { ClusterProvider } from "./providers/cluster";
-import {
-  TransactionsProvider,
-  useTransactionsDispatch,
-  useTransactions,
-  ActionType
-} from "./providers/transactions";
-import {
-  AccountsProvider,
-  useSelectedAccount,
-  useAccountsDispatch,
-  ActionType as AccountActionType
-} from "./providers/accounts";
-import { BlocksProvider } from "./providers/blocks";
 import ClusterStatusButton from "./components/ClusterStatusButton";
 import AccountsCard from "./components/AccountsCard";
 import TransactionsCard from "./components/TransactionsCard";
@@ -28,82 +14,70 @@ function App() {
   const [showClusterModal, setShowClusterModal] = React.useState(false);
   const currentTab = useCurrentTab();
   return (
-    <ClusterProvider>
-      <AccountsProvider>
-        <TransactionsProvider>
-          <BlocksProvider>
-            <ClusterModal
-              show={showClusterModal}
-              onClose={() => setShowClusterModal(false)}
-            />
-            <TransactionModal />
-            <AccountModal />
-            <div className="main-content">
-              <nav className="navbar navbar-expand-xl navbar-light">
-                <div className="container">
-                  <div className="row align-items-end">
-                    <div className="col">
-                      <img src={Logo} width="250" alt="Solana Explorer" />
-                    </div>
-                  </div>
-                </div>
-              </nav>
+    <>
+      <ClusterModal
+        show={showClusterModal}
+        onClose={() => setShowClusterModal(false)}
+      />
+      <TransactionModal />
+      <AccountModal />
+      <div className="main-content">
+        <nav className="navbar navbar-expand-xl navbar-light">
+          <div className="container">
+            <div className="row align-items-end">
+              <div className="col">
+                <img src={Logo} width="250" alt="Solana Explorer" />
+              </div>
+            </div>
+          </div>
+        </nav>
 
-              <div className="header">
-                <div className="container">
-                  <div className="header-body">
-                    <div className="row align-items-center d-md-none">
-                      <div className="col-12">
-                        <ClusterStatusButton
-                          expand
-                          onClick={() => setShowClusterModal(true)}
-                        />
-                      </div>
-                    </div>
-                    <div className="row align-items-center">
-                      <div className="col">
-                        <ul className="nav nav-tabs nav-overflow header-tabs">
-                          <li className="nav-item">
-                            <NavLink href="/transactions" tab="Transactions" />
-                          </li>
-                          <li className="nav-item">
-                            <NavLink href="/accounts" tab="Accounts" />
-                          </li>
-                        </ul>
-                      </div>
-                      <div className="col-auto d-none d-md-block">
-                        <ClusterStatusButton
-                          onClick={() => setShowClusterModal(true)}
-                        />
-                      </div>
-                    </div>
-                  </div>
+        <div className="header">
+          <div className="container">
+            <div className="header-body">
+              <div className="row align-items-center d-md-none">
+                <div className="col-12">
+                  <ClusterStatusButton
+                    expand
+                    onClick={() => setShowClusterModal(true)}
+                  />
                 </div>
               </div>
-
-              <div className="container">
-                <div className="row">
-                  <div className="col-12">
-                    {currentTab === "Transactions" ? (
-                      <TransactionsCard />
-                    ) : null}
-                  </div>
+              <div className="row align-items-center">
+                <div className="col">
+                  <ul className="nav nav-tabs nav-overflow header-tabs">
+                    <li className="nav-item">
+                      <NavLink href="/transactions" tab="Transactions" />
+                    </li>
+                    <li className="nav-item">
+                      <NavLink href="/accounts" tab="Accounts" />
+                    </li>
+                  </ul>
                 </div>
-                <div className="row">
-                  <div className="col-12">
-                    {currentTab === "Accounts" ? <AccountsCard /> : null}
-                  </div>
+                <div className="col-auto d-none d-md-block">
+                  <ClusterStatusButton
+                    onClick={() => setShowClusterModal(true)}
+                  />
                 </div>
               </div>
             </div>
-            <Overlay
-              show={showClusterModal}
-              onClick={() => setShowClusterModal(false)}
-            />
-          </BlocksProvider>
-        </TransactionsProvider>
-      </AccountsProvider>
-    </ClusterProvider>
+          </div>
+        </div>
+
+        <div className="container">
+          <div className="row">
+            <div className="col-12">
+              {currentTab === "Transactions" ? <TransactionsCard /> : null}
+            </div>
+          </div>
+          <div className="row">
+            <div className="col-12">
+              {currentTab === "Accounts" ? <AccountsCard /> : null}
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
   );
 }
 
@@ -118,32 +92,6 @@ function NavLink({ href, tab }: { href: string; tab: Tab }) {
       {tab}
     </Link>
   );
-}
-
-type OverlayProps = {
-  show: boolean;
-  onClick: () => void;
-};
-
-function Overlay({ show, onClick }: OverlayProps) {
-  const { selected } = useTransactions();
-  const selectedAccount = useSelectedAccount();
-  const txDispatch = useTransactionsDispatch();
-  const accountDispatch = useAccountsDispatch();
-
-  if (show || !!selected || !!selectedAccount)
-    return (
-      <div
-        className="modal-backdrop fade show"
-        onClick={() => {
-          onClick();
-          txDispatch({ type: ActionType.Deselect });
-          accountDispatch({ type: AccountActionType.Select });
-        }}
-      ></div>
-    );
-
-  return <div className="fade"></div>;
 }
 
 export default App;
